refactor(wu-tuner): move types to separate modules
