+++ wu-tuner/src/lib.rs
@@ -2,7 +2,8 @@ pub use config::WuTunerConfig;
 use tycho_collator::collator::work_units::{MempoolAnchorLag, WuEvent, WuEventData, WuMetrics};
 
 mod config;
-pub(crate) mod unit_cost_clipper;
 pub mod service;
 mod tuner;
+mod types;
+mod unit_cost_clipper;
 pub mod updater;
