+++ wu-tuner/src/types/mod.rs
@@ -1,0 +1,7 @@
+mod do_collate_wu_avg;
+mod execute_wu_avg;
+mod finalize_wu_avg;
+mod prepare_wu_avg;
+mod wu_metrics_avg;
+
+pub use wu_metrics_avg::WuMetricsAvg;
